Another x64 fix
git-svn-id: https://code.ros.org/svn/ros-pkg/pkg/trunk/stacks/visualization@22581 eb33c2ac-9c88-4c90-87e0-44a10359b0c3

+++ src/rviz/plugin_manager_dialog.cpp
@@ -116,7 +116,7 @@ void PluginManagerDialog::onLoadedChecked(wxCommandEvent& evt)
 void PluginManagerDialog::onAutoLoadChecked(wxCommandEvent& evt)
 {
   wxCheckBox* box = (wxCheckBox*)evt.GetEventObject();
-  uint32_t row_index = (uint32_t)box->GetClientData();
+  uintptr_t row_index = (uintptr_t)box->GetClientData();
   Plugin* plugin = rows_[row_index].plugin;
 
   plugin->setAutoLoad(evt.IsChecked());
